requirements.txt
Pillow==10.1.0
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
shiboken6==6.6.0
